Add files via upload
Finite geometry has numerous applications in networking and computer science. 
In this presentation, we explore one application to the Hamming code, a type of
error-correction coding widely used for transmitting data over noisy networks and 
for preserving the integrity of data in memory or storage for essential applications. 
We focus here in particular on Fano’s geometry, also called the seven-point geometry 
or PG(2,2).